2009-11-05  Ivan Maidanski <ivmai@mail.ru>
	* dyn_load.c (GC_dump_meminfo): Prefix "%lx" printf type specifier
	with "0x".
	* os_dep.c (PROTECT): Ditto.
	* win32_threads.c (GC_mark_thread_local_free_lists): Cast p->id to
	int (to match printf type specifier).

+++ win32_threads.c
@@ -2635,7 +2635,7 @@ GC_INNER void GC_init_parallel(void)
     for (i = 0; i < THREAD_TABLE_SZ; ++i) {
       for (p = GC_threads[i]; 0 != p; p = p -> tm.next) {
 #       ifdef DEBUG_THREADS
-          GC_printf("Marking thread locals for 0x%x\n", p -> id);
+          GC_printf("Marking thread locals for 0x%x\n", (int)p -> id);
 #       endif
         GC_mark_thread_local_fls_for(&(p->tlfs));
       }
